requirements/base.txt
logzero
tqdm
numpy
vcfpy
pysam
pandas
plotly
scipy
attrs
cattrs
